+++ capture/src/api.rs
@@ -1,8 +1,11 @@
-use crate::token::InvalidTokenReason;
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
+use time::OffsetDateTime;
+use uuid::Uuid;
+
+use crate::token::InvalidTokenReason;
 
 #[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
 pub enum CaptureResponseCode {
@@ -71,5 +74,23 @@ impl IntoResponse for CaptureError {
             }
         }
         .into_response()
+    }
+}
+
+#[derive(Clone, Default, Debug, Serialize, Eq, PartialEq)]
+pub struct ProcessedEvent {
+    pub uuid: Uuid,
+    pub distinct_id: String,
+    pub ip: String,
+    pub data: String,
+    pub now: String,
+    #[serde(with = "time::serde::rfc3339::option")]
+    pub sent_at: Option<OffsetDateTime>,
+    pub token: String,
+}
+
+impl ProcessedEvent {
+    pub fn key(&self) -> String {
+        format!("{}:{}", self.token, self.distinct_id)
     }
 }
