+++ capture/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod api;
-pub mod capture;
 pub mod config;
-pub mod event;
 pub mod health;
 pub mod limiters;
 pub mod prometheus;
@@ -12,3 +10,5 @@ pub mod sinks;
 pub mod time;
 pub mod token;
 pub mod utils;
+pub mod v0_endpoint;
+pub mod v0_request;
