+++ capture/tests/django_compat.rs
@@ -4,8 +4,7 @@ use axum::http::StatusCode;
 use axum_test_helper::TestClient;
 use base64::engine::general_purpose;
 use base64::Engine;
-use capture::api::{CaptureError, CaptureResponse, CaptureResponseCode};
-use capture::event::ProcessedEvent;
+use capture::api::{CaptureError, CaptureResponse, CaptureResponseCode, ProcessedEvent};
 use capture::health::HealthRegistry;
 use capture::limiters::billing::BillingLimiter;
 use capture::redis::MockRedisClient;
